requirements.txt
python-docx==0.8.11
google-generativeai==0.8.3
python-dotenv==1.0.1
colorama==0.4.6
tqdm==4.66.5
lxml==6.0.0
rapidfuzz==3.6.1
psutil==5.9.8